Reorg staking related code into staking.go
